feat(core): update for better any handling and imports
Going full-TypeScript, we're including no-any for linting and no-implicit-any for configuration.

This also includes a new tslint.fix.json that has fixable rules (currently just ordering and grouping imports) that are not handled by prettier. This will be used for commit hooks.

Finally, the documentation was pretty out-of-date with more recent changes. I've updated it to reflect all current changes.